requirements.txt
python-dotenv
huggingface_hub
requests